Added final mutations and queries on back end. Added admin commands for the front end and fixed some bugs
