+++ client/package.json
@@ -4,21 +4,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.8.3",
-    "@material-ui/icons": "^4.5.1",
+    "@apollo/react-hooks": "^3.1.3",
+    "@material-ui/core": "^4.9.1",
+    "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.39",
     "@react-spring/web": "^9.0.0-beta.34",
     "apollo-boost": "^0.4.7",
-    "graphql": "^14.5.8",
+    "graphql": "^14.6.0",
     "graphql-tag": "^2.10.1",
+    "js-cookie": "^2.2.1",
     "moment": "^2.24.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-google-login": "^5.0.7",
-    "react-scripts": "3.3.0",
+    "react-google-login": "^5.1.0",
+    "react-scripts": "3.3.1",
     "react-spring": "^8.0.27",
     "styled-components": "^5.0.0",
-    "victory": "^34.0.0"
+    "victory": "^34.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -40,5 +42,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "devDependencies": {}
 }
