+++ client/src/utils.js
@@ -1,38 +1,50 @@
 import ApolloClient from 'apollo-boost';
+import React from 'react';
+import { ApolloProvider } from '@apollo/react-hooks';
 
 const client = new ApolloClient({
-    // uri: 'https://volence.dev/node_apps/lighthouse_app/graphql',
-    uri: 'http://localhost:3501//graphql',
+    uri: 'https://volence.dev/node_apps/lighthouse_app/graphql',
+    // uri: 'http://localhost:3501//graphql',
     // 100.11.206.213
 });
 
-const sendQuery = query => {
-    return new Promise(async (res, rej) => {
-        try {
-            let response = await client.query({
-                query: query,
-            });
-            res(response);
-        } catch (err) {
-            rej(err);
-        }
-    });
+const sendQuery = async query => {
+    try {
+        let response = await client.query({
+            query: query,
+        });
+        return response;
+    } catch (err) {
+        return err;
+    }
 };
-const sendMutation = (mutationQuery, mutationVariable) => {
-    console.log('mutationVariable ', mutationVariable);
-    return new Promise(async (res, rej) => {
-        try {
-            let response = await client.mutate({
-                mutation: mutationQuery,
-                variables: {
-                    siteInput: mutationVariable,
-                },
-            });
-            res(response);
-        } catch (err) {
-            rej(err);
-        }
-    });
+const sendMutation = async (mutationQuery, mutationVariable) => {
+    try {
+        let response = await client.mutate({
+            mutation: mutationQuery,
+            variables: {
+                siteInput: mutationVariable,
+            },
+        });
+        return response;
+    } catch (err) {
+        return err;
+    }
+};
+const sendSignInMutation = async (mutationQuery, mutationVariable) => {
+    try {
+        let response = await client.mutate({
+            mutation: mutationQuery,
+            variables: {
+                clientData: mutationVariable,
+            },
+        });
+        return response;
+    } catch (err) {
+        return err;
+    }
 };
 
-export { sendQuery, sendMutation };
+const AppoloWrapper = ({ children }) => <ApolloProvider client={client}>{children}</ApolloProvider>;
+
+export { sendQuery, sendMutation, sendSignInMutation, AppoloWrapper };
