+++ server/.directory
@@ -1,0 +1,6 @@
+[Dolphin]
+Timestamp=2020,2,2,3,5,26
+Version=4
+
+[Settings]
+HiddenFilesShown=true
