+++ server/package.json
@@ -13,11 +13,15 @@
     "cron": "^1.8.2",
     "express": "^4.17.1",
     "express-graphql": "^0.9.0",
+    "google-auth-library": "^5.9.2",
     "graphql": "^14.6.0",
     "graphql-tools": "^4.0.6",
+    "jsonwebtoken": "^8.5.1",
     "lighthouse": "^5.6.0",
     "moment": "^2.24.0",
     "mongoose": "^5.8.10",
+    "passport": "^0.4.1",
+    "passport-google-oauth2": "^0.2.0",
     "puppeteer": "^2.1.0"
   },
   "scripts": {
