+++ server/types/resolvers.d.ts
@@ -8,6 +8,20 @@ interface Arguments {
     };
 }
 
+interface SignInArguments {
+    clientData: {
+        userName?: string;
+        email?: string;
+        clientToken?: string;
+        userID?: string;
+        error?: string;
+    };
+}
+
+interface AdminEditArgs {
+    email: string;
+}
+
 interface Element {}
 interface Node {}
 interface NodeListOf<TNode = Node> {}
